🔨 Refactor: 공통컴포넌트 import index 사용
