+++ src/app/[teamid]/_components/member/card.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 
-import DropDown from "@/components/common/drop-down";
+import { DropDown } from "@/components/common";
 import useToggle from "@/hooks/use-toggle";
 import { IconProfile } from "@/public/assets/icons";
